# console/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/connections/test": {
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Test Connection",
        "description": "Dial a connection before saving it, so a typo is caught here rather than\nat the first query.",
        "operationId": "test_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionProbeModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connections": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "List Connections",
        "operationId": "list_connections",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsModelList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Create Connection",
        "operationId": "create_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionsModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connections/{connection_uid}": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "Get Connection",
        "operationId": "get_connection",
        "parameters": [
          {
            "name": "connection_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "connections"
        ],
        "summary": "Update Connection",
        "operationId": "update_connection",
        "parameters": [
          {
            "name": "connection_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConnectionsModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "connections"
        ],
        "summary": "Delete Connection",
        "operationId": "delete_connection",
        "parameters": [
          {
            "name": "connection_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Uid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connections/{connection_uid}/status": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "Get Connection Status",
        "description": "Dial the connection and report whether it answers, without running a query.",
        "operationId": "get_connection_status",
        "parameters": [
          {
            "name": "connection_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionStatusModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionProbeModel": {
        "properties": {
          "reachable": {
            "type": "boolean",
            "title": "Reachable"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "server_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Version"
          }
        },
        "type": "object",
        "required": [
          "reachable"
        ],
        "title": "ConnectionProbeModel"
      },
      "ConnectionRole": {
        "type": "string",
        "enum": [
          "primary",
          "replica"
        ],
        "title": "ConnectionRole"
      },
      "ConnectionStatusModel": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "reachable": {
            "type": "boolean",
            "title": "Reachable"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "server_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Version"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "reachable"
        ],
        "title": "ConnectionStatusModel",
        "description": "Result of dialling a connection without running a user query."
      },
      "ConnectionsModel": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "connection_uri": {
            "type": "string",
            "title": "Connection Uri"
          },
          "source": {
            "$ref": "#/components/schemas/SourceConfig"
          },
          "environment": {
            "$ref": "#/components/schemas/Environment",
            "default": "local"
          },
          "role": {
            "$ref": "#/components/schemas/ConnectionRole",
            "default": "primary"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          },
          "supports_schemas": {
            "type": "boolean",
            "title": "Supports Schemas",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "uid",
          "name",
          "connection_uri",
          "source"
        ],
        "title": "ConnectionsModel"
      },
      "ConnectionsModelList": {
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ConnectionsModel"
            },
            "type": "array",
            "title": "Connections"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "connections",
          "total"
        ],
        "title": "ConnectionsModelList"
      },
      "CreateConnectionsModel": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/SourceConfig"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "connection_uri": {
            "type": "string",
            "title": "Connection Uri"
          },
          "environment": {
            "$ref": "#/components/schemas/Environment",
            "default": "local"
          },
          "role": {
            "$ref": "#/components/schemas/ConnectionRole",
            "default": "primary"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "source",
          "name",
          "connection_uri"
        ],
        "title": "CreateConnectionsModel"
      },
      "Environment": {
        "type": "string",
        "enum": [
          "local",
          "staging",
          "production"
        ],
        "title": "Environment",
        "description": "Where a connection points, so the UI can warn before touching production."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceConfig": {
        "type": "string",
        "enum": [
          "postgres",
          "sqlite",
          "mysql",
          "api"
        ],
        "title": "SourceConfig"
      },
...
      "TestConnectionModel": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/SourceConfig"
          },
          "connection_uri": {
            "type": "string",
            "title": "Connection Uri"
          }
        },
        "type": "object",
        "required": [
          "source",
          "connection_uri"
        ],
        "title": "TestConnectionModel",
        "description": "Enough to dial a connection that has not been saved yet."
      },
      "UpdateConnectionsModel": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "connection_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Uri"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "environment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Environment"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectionRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "read_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read Only"
          }
        },
        "type": "object",
        "title": "UpdateConnectionsModel"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}